+++ src/gdal_algs.cpp
@@ -806,7 +806,7 @@ SpatVector SpatRaster::polygonize(bool trunc, bool values, bool narm, bool aggre
 
 	std::vector<double> fext;
 	SpatVector fvct;
-	out.read_ogr(poDS, "", "", fext, fvct);
+	out.read_ogr(poDS, "", "", fext, fvct, false);
 	GDALClose(poDS);
 
 	if (aggregate && (out.nrow() > 0)) {