+++ src/read_gdal.cpp
@@ -229,7 +229,7 @@ bool GetVAT(std::string filename, SpatCategories &vat) {
 	SpatVector v, fvct;
 	std::vector<double> fext;
 
-	v.read(filename, "", "", fext, fvct);
+	v.read(filename, "", "", fext, fvct, true);
 	if (v.df.nrow() == 0) return false;