+++ src/vector_methods.cpp
@@ -60,7 +60,7 @@ SpatVector SpatVector::make_valid() {
 	}
 	std::vector<double> fext;
 	SpatVector fvct;
-	out.read_ogr(src, "", "", fext, fvct);
+	out.read_ogr(src, "", "", fext, fvct, false);
 	GDALClose(src);
 	return out;
 }